Exclusive FullScreen Bug Documented
Known bug: If you ever lose focus of the game (alt tabing, task manager) while in exclusive full screen and close the game (end task, close window, etc.), the fullscreen mode registry will save it as a value of 3 (windowed) instead of 0 (exclusive full screen).

Therefore, when you reopen the game, it returns to windowed mode.
If you leave the game through the in game button, this doesn't happen.
I do not know why this interaction occurs.
This is partially solved with binary file saving.